Updated preview music handling fallback to 40% of song length

+++ src/App/Osu/OsuDatabaseBeatmap.h
@@ -171,7 +171,7 @@ public:
 	inline const UString &getAudioFileName() const {return m_sAudioFileName;}
 
 	inline unsigned long getLengthMS() const {return m_iLengthMS;}
-	inline unsigned long getPreviewTime() const {return m_iPreviewTime;}
+	inline int getPreviewTime() const {return m_iPreviewTime;}
 
 	inline float getAR() const {return m_fAR;}
 	inline float getCS() const {return m_fCS;}
@@ -235,7 +235,7 @@ private:
 	UString m_sAudioFileName;
 
 	unsigned long m_iLengthMS;
-	unsigned long m_iPreviewTime;
+	int m_iPreviewTime;
 
 	float m_fAR;
 	float m_fCS;
